feat:create inviteService

+++ src/services/user/userDelete.service.ts
@@ -1,4 +1,11 @@
-export const deleteService = async(id: String): Promise<void> => {
-  const user = await userRepository.findOneBy({id})
+import User from "../../entities/user.entity"
+import { AppError } from "../../errors/errors"
+import { userRepository } from "../../repositories/userRepository"
+
+export const deleteService = async(id: string): Promise<User> => {
+  const user = await userRepository.findOneBy({id: id})
+  if(!user){
+    throw new AppError(404, 'user not found')
+  } 
   return await userRepository.remove(user)
 }
